Add a set of NMake Makefiles
This adds a set of NMake Makefiles with pre-configured config.h template for
building dbus-python using Visual Studio, supporting builds against the
Python versions that are supported by dbus-python itself.  This will build
both the DBus and DBus-GLib Python bindings against the Python interpreter that
is specified, or the one that is found in %PATH%, and supports packaging
preparation via using ZIP files afterwards or in a later patch, DistUtils and
Wheels.